NestedWorkflowSpans working with updated state and selectors for new data structures.
