+++ frontend/src/features/junjo-data/workflow-detail/WorkflowDetailStateDiff.tsx
@@ -6,16 +6,16 @@ import { detailedDiff, diff } from 'deep-object-diff'
 import { TriangleDownIcon } from '@radix-ui/react-icons'
 import { useAppSelector } from '../../../root-store/hooks'
 import { RootState } from '../../../root-store/store'
-import {
-  selectActiveSpansWorkflowSpan,
-  selectActiveStoreID,
-  selectBeforeSpanStateEventInWorkflow,
-  selectSetStateEventsByStoreID,
-  selectWorkflowSpanByStoreID,
-} from '../../otel/store/selectors'
+import { selectWorkflowSpanByStoreID } from '../../otel/store/selectors'
 import * as jsonpatch from 'fast-json-patch'
 import { OtelSpan } from '../../otel/schemas/schemas'
 import SpanExceptionsList from './SpanExceptionsList'
+import {
+  selectActiveSpanJunjoWorkflow,
+  selectActiveStoreID,
+  selectBeforeSpanStateEventInWorkflow,
+  selectStateEventsByJunjoStoreId,
+} from '../../traces/store/selectors'
 
 enum DiffTabOptions {
   BEFORE = 'Before',
@@ -77,24 +77,13 @@ export default function WorkflowDetailStateDiff(props: WorkflowDetailStateDiffPr
 
   // Get All Workflow State Events
   // This includes the default top level workflow and all subflows
-  const defaultWorkflowSelectorProps = useMemo(
-    () => ({
-      serviceName,
-      spanID: defaultWorkflowSpanID,
-    }),
-    [serviceName, defaultWorkflowSpanID],
-  )
   // The active workflow span is the workflow with store.id matching the activeStoreID,
   // defaulting to the default workflow span if there is no active set state event
   // This may be a subflow of the default workflow depending on the set state event
-  const activeWorkflowSpan = useAppSelector((state: RootState) =>
-    selectActiveSpansWorkflowSpan(state, defaultWorkflowSelectorProps),
-  )
+  const activeWorkflowSpan = useAppSelector((state: RootState) => selectActiveSpanJunjoWorkflow(state))
 
   // The Active Store is the store that the active span is acting on
-  const activeStoreID = useAppSelector((state: RootState) =>
-    selectActiveStoreID(state, defaultWorkflowSelectorProps),
-  )
+  const activeStoreID = useAppSelector((state: RootState) => selectActiveStoreID(state))
 
   // This is the workflow span that owns the store that the active span is acting on
   // The active span may be a subflow operating on a parent store.
@@ -113,16 +102,12 @@ export default function WorkflowDetailStateDiff(props: WorkflowDetailStateDiffPr
   // This is only state events for the currently actively rendering workflow or subflow
   // This is so we can construct the patches of just this rendered JSON state for this workflow and its store
   // This will not include state events of parent or child stores
-  const activeWorkflowSelectorProps = useMemo(
-    () => ({
-      serviceName,
-      spanID: defaultWorkflowSpanID,
-      storeID: activeStoreID,
-    }),
-    [serviceName, defaultWorkflowSpanID, activeStoreID],
-  )
   const activeStoreStateEvents = useAppSelector((state: RootState) =>
-    selectSetStateEventsByStoreID(state, activeWorkflowSelectorProps),
+    selectStateEventsByJunjoStoreId(state, {
+      traceId: activeWorkflowSpan?.trace_id,
+      spanId: activeWorkflowSpan?.span_id,
+      storeId: activeStoreID,
+    }),
   )
 
   // The starting state of the active workflow
@@ -136,7 +121,11 @@ export default function WorkflowDetailStateDiff(props: WorkflowDetailStateDiffPr
 
   // Select: Gets the last set_state event before the active span
   const beforeActiveSpanStateEvent = useAppSelector((state: RootState) =>
-    selectBeforeSpanStateEventInWorkflow(state, activeWorkflowSelectorProps),
+    selectBeforeSpanStateEventInWorkflow(state, {
+      traceId: activeWorkflowSpan?.trace_id,
+      spanId: activeWorkflowSpan?.span_id,
+      storeId: activeStoreID,
+    }),
   )
 
   // Local State
@@ -183,82 +172,82 @@ export default function WorkflowDetailStateDiff(props: WorkflowDetailStateDiffPr
     }
   }, [activeTab, hasExceptions])
 
-  /**
-   * Accumulate State Patches To Index (inclusive)
-   *
-   * Given a patch index, this function will accumulate the patches up to and including the patch at the given index.
-   *
-   * @returns {[Record<string, any>, Record<string, any>]} - before / after state
-   */
-  const accumulateStatePathesToIndex = (patchIndex: number): [Record<string, any>, Record<string, any>] => {
-    // If there are no patches, just set the original state
-    if (activeStoreStateEvents.length === 0) {
-      return [workflowStateStart, workflowStateStart]
-    }
+  // /**
+  //  * Accumulate State Patches To Index (inclusive)
+  //  *
+  //  * Given a patch index, this function will accumulate the patches up to and including the patch at the given index.
+  //  *
+  //  * @returns {[Record<string, any>, Record<string, any>]} - before / after state
+  //  */
+  // const accumulateStatePathesToIndex = (patchIndex: number): [Record<string, any>, Record<string, any>] => {
+  //   // If there are no patches, just set the original state
+  //   if (activeStoreStateEvents.length === 0) {
+  //     return [workflowStateStart, workflowStateStart]
+  //   }
 
-    // If the patch index is out of bounds, return (there is no patch)
-    if (patchIndex < 0 || patchIndex >= activeStoreStateEvents.length) {
-      return [workflowStateStart, workflowStateStart]
-    }
+  //   // If the patch index is out of bounds, return (there is no patch)
+  //   if (patchIndex < 0 || patchIndex >= activeStoreStateEvents.length) {
+  //     return [workflowStateStart, workflowStateStart]
+  //   }
 
-    // Starting points for accumulating patches
-    let beforeCumulativeState = structuredClone(workflowStateStart)
-    let afterCumulativeState = structuredClone(workflowStateStart)
+  //   // Starting points for accumulating patches
+  //   let beforeCumulativeState = structuredClone(workflowStateStart)
+  //   let afterCumulativeState = structuredClone(workflowStateStart)
 
-    // Apply patches to the cumulative state
-    for (let i = 0; i <= patchIndex; i++) {
-      const thisEvent = activeStoreStateEvents[i]
+  //   // Apply patches to the cumulative state
+  //   for (let i = 0; i <= patchIndex; i++) {
+  //     const thisEvent = activeStoreStateEvents[i]
 
-      const patchString = thisEvent.attributes['junjo.state_json_patch']
-      const patch = JSON.parse(patchString)
-      // console.log(`Patch ${i} of ${patchIndex}`)
-      // console.log('Patch string: ', patchString)
-      // console.log('Patch: ', patch)
+  //     const patchString = thisEvent.attributes['junjo.state_json_patch']
+  //     const patch = JSON.parse(patchString)
+  //     // console.log(`Patch ${i} of ${patchIndex}`)
+  //     // console.log('Patch string: ', patchString)
+  //     // console.log('Patch: ', patch)
 
-      // Apply to after state
-      afterCumulativeState = jsonpatch.applyPatch(afterCumulativeState, patch).newDocument
+  //     // Apply to after state
+  //     afterCumulativeState = jsonpatch.applyPatch(afterCumulativeState, patch).newDocument
 
-      // Apply to before state if i is less than patchIndex
-      if (i < patchIndex) {
-        beforeCumulativeState = jsonpatch.applyPatch(beforeCumulativeState, patch).newDocument
-      }
-    }
+  //     // Apply to before state if i is less than patchIndex
+  //     if (i < patchIndex) {
+  //       beforeCumulativeState = jsonpatch.applyPatch(beforeCumulativeState, patch).newDocument
+  //     }
+  //   }
 
-    return [beforeCumulativeState, afterCumulativeState]
-  }
+  //   return [beforeCumulativeState, afterCumulativeState]
+  // }
 
-  /**
-   * Run the patch accumulation functions based on the active span / active state event
-   */
-  useEffect(() => {
-    // If there is no active set state event,
-    // use the index of the set_state event that occurs most recently before
-    // the active span start time
-    // NOTE: THIS DEPENDS ON THE AUTOMATIC SELECTION OF THE FIRST STATE EVENT INSIDE SPANS WITH STATE EVENTS
-    //       OTHERWISE, the ux diffs may not make sense.
-    //       Spans with no state events have the same before / after state (equal to the AFTER state of the most recent prior state event)
-    if (!activeSetStateEvent) {
-      // The index of the set state event that occurs just prior to the active span, in the list of state events for the active workflow's store
-      const indexOfBeforeActiveSpanSetStateEventInsideActiveStore = activeStoreStateEvents.findIndex(
-        (event) => event.attributes.id === beforeActiveSpanStateEvent?.attributes.id,
-      )
-      const [_before, after] = accumulateStatePathesToIndex(
-        indexOfBeforeActiveSpanSetStateEventInsideActiveStore,
-      )
-      setBeforeJson(after)
-      setAfterJson(after)
-      return
-    }
+  // /**
+  //  * Run the patch accumulation functions based on the active span / active state event
+  //  */
+  // useEffect(() => {
+  //   // If there is no active set state event,
+  //   // use the index of the set_state event that occurs most recently before
+  //   // the active span start time
+  //   // NOTE: THIS DEPENDS ON THE AUTOMATIC SELECTION OF THE FIRST STATE EVENT INSIDE SPANS WITH STATE EVENTS
+  //   //       OTHERWISE, the ux diffs may not make sense.
+  //   //       Spans with no state events have the same before / after state (equal to the AFTER state of the most recent prior state event)
+  //   if (!activeSetStateEvent) {
+  //     // The index of the set state event that occurs just prior to the active span, in the list of state events for the active workflow's store
+  //     const indexOfBeforeActiveSpanSetStateEventInsideActiveStore = activeStoreStateEvents.findIndex(
+  //       (event) => event.attributes.id === beforeActiveSpanStateEvent?.attributes.id,
+  //     )
+  //     const [_before, after] = accumulateStatePathesToIndex(
+  //       indexOfBeforeActiveSpanSetStateEventInsideActiveStore,
+  //     )
+  //     setBeforeJson(after)
+  //     setAfterJson(after)
+  //     return
+  //   }
 
-    // If there is an active set state event, use the index of that event
-    // The before / after is based on the before patch state, and after patch state
-    const indexOfActiveSetStateEventInsideActiveStore = activeStoreStateEvents.findIndex(
-      (event) => event.attributes.id === activeSetStateEvent?.attributes.id,
-    )
-    const [before, after] = accumulateStatePathesToIndex(indexOfActiveSetStateEventInsideActiveStore)
-    setBeforeJson(before)
-    setAfterJson(after)
-  }, [activeStoreStateEvents, activeWorkflowSpan, activeSetStateEvent, beforeActiveSpanStateEvent])
+  //   // If there is an active set state event, use the index of that event
+  //   // The before / after is based on the before patch state, and after patch state
+  //   const indexOfActiveSetStateEventInsideActiveStore = activeStoreStateEvents.findIndex(
+  //     (event) => event.attributes.id === activeSetStateEvent?.attributes.id,
+  //   )
+  //   const [before, after] = accumulateStatePathesToIndex(indexOfActiveSetStateEventInsideActiveStore)
+  //   setBeforeJson(before)
+  //   setAfterJson(after)
+  // }, [activeStoreStateEvents, activeWorkflowSpan, activeSetStateEvent, beforeActiveSpanStateEvent])
 
   // Get Tab Collapsed Level
   const getTabCollapsedLevel = (tab: DiffTabOptions) => {
